Created metadata and callbacks class for runner and reorged code
